# ui/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Artificial Planeswalker Companion",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/cards/{card_id}": {
      "get": {
        "summary": "Read Card",
        "description": "Return everything known about one card printing.\n\nThe canonical record behind a printing id: its name, mana cost, converted mana cost, type\nline, oracle text, power and toughness, rarity, set, collector number, colours, keywords,\nformat legalities, and its images. Views that were given only an id use this to fill\nthemselves in.\n\nThis operation returns the image **URLs** and nothing else about them; where they live on the\nrecord is stated on the ``image_uris`` and ``card_faces`` fields themselves, and it is not the\nobvious rule. A view that wants pixels asks ``GET /api/card-image/{scryfall_id}`` instead,\nwhich applies that rule for you.\n\n``prices`` is absent from this response, not empty: the local database holds no price data of\nany kind.\n\nWarning:\n    The ``400`` for a malformed id is not unconditional. Dependencies resolve before\n    parameter validation is reported, so a malformed id sent while the database is unusable\n    answers ``503`` (``database_not_initialized`` or ``database_unavailable``), not ``400``.\n    A caller that retries on 503 must not assume the request can ever succeed: a malformed id\n    stays malformed whatever the database does.",
        "operationId": "read_card_api_cards__card_id__get",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Card Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Card"
                }
              }
            }
          },
          "400": {
            "description": "reason: invalid_request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "reason: payload_too_large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "reason: database_unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "reason: internal_error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "reason: card_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/card-image/{scryfall_id}": {
      "get": {
        "summary": "Read Card Image",
        "description": "Serve one card face's artwork, from this app's own origin.\n\nEvery image in the app comes through here: the browser never contacts Scryfall directly\n(AD-11, UX-DR36). The URL is read from the card's own row in the local database — **no live\nScryfall metadata call is ever made** — and the bytes are proxied back with the content type\nthe CDN actually sent.\n\n``face`` indexes **the images this card has**, not its ``card_faces`` array, and the two are\ndifferent for real cards. A split, adventure or flip card has two faces and **one** image,\nbecause its halves share one piece of artwork, so ``face=1`` on one of them is out of range\nrather than \"the other half\"; a single-faced card serves at ``face=0``. Which images a card\nhas is decided by the rule stated on ``GET /api/cards/{card_id}``'s ``image_uris`` and\n``card_faces`` fields — this operation applies it so a caller does not have to.\n\nTwo failures are answered here and they are **not** interchangeable, even though both draw the\nsame placeholder. ``404 no_image_data`` is permanent: this card has no artwork for what was\nasked, and asking again cannot change it. ``502 image_fetch_failed`` is transient: the URL was\nknown and the fetch did not deliver. Neither ever returns a substitute image — no grey\nrectangle, no 1×1 pixel, no generic card back — because the client can draw a better one from\nthe name, cost and type line it already has.\n\nA successful image is cacheable by the browser for a year; a failure is not cached at all, so\none bad minute cannot leave a permanently broken tile in an open tab.\n\nWarning:\n    The ``400`` for a malformed id or an unrecognised ``size`` is not unconditional.\n    Dependencies resolve before parameter validation is reported, so ``?size=bogus`` sent\n    while the database is unusable answers ``503`` (``database_not_initialized`` or\n    ``database_unavailable``), not ``400``. A caller that retries on 503 must not assume the\n    request can ever succeed: a bad parameter stays bad whatever the database does.",
        "operationId": "read_card_image_api_card_image__scryfall_id__get",
        "parameters": [
          {
            "name": "scryfall_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "title": "Scryfall Id"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "small",
                "normal",
                "large",
                "png",
                "art_crop",
                "border_crop"
              ],
              "type": "string",
              "default": "normal",
              "title": "Size"
            }
          },
          {
            "name": "face",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Face"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The image bytes, in whatever format the CDN served them.",
            "content": {
              "image/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "reason: invalid_request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "reason: payload_too_large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "reason: database_unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "reason: internal_error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "reason: card_not_found | no_image_data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "reason: image_fetch_failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Card": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "printed_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Printed Name"
          },
          "oracle_id": {
            "type": "string",
            "title": "Oracle Id"
          },
          "mana_cost": {
            "type": "string",
            "title": "Mana Cost"
          },
          "cmc": {
            "type": "number",
            "title": "Cmc"
          },
          "type_line": {
            "type": "string",
            "title": "Type Line"
          },
          "oracle_text": {
            "type": "string",
            "title": "Oracle Text"
          },
          "power": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Power"
          },
          "toughness": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toughness"
          },
          "game_changer": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Game Changer"
          },
          "rarity": {
            "type": "string",
            "title": "Rarity"
          },
          "set_code": {
            "type": "string",
            "title": "Set Code"
          },
          "set_name": {
            "type": "string",
            "title": "Set Name"
          },
          "collector_number": {
            "type": "string",
            "title": "Collector Number"
          },
          "colors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Colors"
          },
          "color_identity": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Color Identity"
          },
          "color_indicator": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color Indicator"
          },
          "keywords": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keywords"
          },
          "legalities": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Legalities"
          },
          "card_faces": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CardFace"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Card Faces",
            "description": "Images live in one of two places, and which one is decided by the presence of per-face image_uris — never by whether card_faces is present. Most cards carry a top-level image_uris; a card whose faces have their own artwork carries a null image_uris and per-face image_uris inside its card_faces entries instead. The two are mutually exclusive; nothing carries both. Reading 'has faces' as 'has per-face images' is the trap: a split card has faces and a top-level image, because its halves share one piece of artwork, so its faces carry names and costs but no images of their own. Some cards have no image data anywhere, which is ordinary and not an error."
          },
          "image_uris": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Uris",
            "description": "Images live in one of two places, and which one is decided by the presence of per-face image_uris — never by whether card_faces is present. Most cards carry a top-level image_uris; a card whose faces have their own artwork carries a null image_uris and per-face image_uris inside its card_faces entries instead. The two are mutually exclusive; nothing carries both. Reading 'has faces' as 'has per-face images' is the trap: a split card has faces and a top-level image, because its halves share one piece of artwork, so its faces carry names and costs but no images of their own. Some cards have no image data anywhere, which is ordinary and not an error."
          },
          "games": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Games",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "oracle_id",
          "mana_cost",
          "cmc",
          "type_line",
          "oracle_text",
          "rarity",
          "set_code",
          "set_name",
          "collector_number",
          "colors",
          "color_identity",
          "legalities"
        ],
        "title": "Card",
        "description": "One Magic: The Gathering card printing, as held in the local card database.\n\nEverything known about a single printing: its name and mana cost, type line and oracle text,\npower and toughness, rarity, set and collector number, colours, keywords, format legalities,\nand its images.\n\nSome fields are always present but may be empty rather than absent — ``mana_cost`` and\n``oracle_text`` are empty strings on a card that has none, ``colors`` and ``games`` empty\nlists, ``legalities`` an empty object. Combat stats are null on anything that is not a\ncreature, and ``game_changer`` is a three-state flag whose null means \"not yet determined\",\nnot \"no\".\n\nWhere a card's artwork lives is stated on the ``image_uris`` and ``card_faces`` fields\nthemselves — read those descriptions before rendering anything, because the rule is not the\nobvious one.\n\nThere is no price data of any kind in this record."
      },
      "CardFace": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "mana_cost": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mana Cost"
          },
          "type_line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type Line"
          },
          "oracle_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oracle Text"
          },
          "image_uris": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Uris",
            "description": "Images live in one of two places, and which one is decided by the presence of per-face image_uris — never by whether card_faces is present. Most cards carry a top-level image_uris; a card whose faces have their own artwork carries a null image_uris and per-face image_uris inside its card_faces entries instead. The two are mutually exclusive; nothing carries both. Reading 'has faces' as 'has per-face images' is the trap: a split card has faces and a top-level image, because its halves share one piece of artwork, so its faces carry names and costs but no images of their own. Some cards have no image data anywhere, which is ordinary and not an error."
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "CardFace",
        "description": "One face of a multi-faced card printing, exactly as Scryfall publishes it.\n\nA card's ``card_faces`` entries carry each half of its identity — that face's own name, mana\ncost, type line and oracle text — and, on some cards, its own artwork. Five fields are named\nbelow; **every other key Scryfall sends is kept and served unchanged**, so a consumer that\nknows about ``power``, ``flavor_text``, ``loyalty``, ``defense``, ``artist``, ``printed_name``\nor ``color_indicator`` still finds them.\n\nA card having faces does **not** mean its faces have artwork. The ``image_uris`` field below\ncarries the rule that decides it; this summary deliberately does not repeat it."
      },
...
    }
  }
}